/Akshay

+++ src/main/java/com/ats/tril/model/report/IssueReportDetail.java
@@ -25,11 +25,11 @@ public class IssueReportDetail {
 
 	private int itemId;
 
-	private int itemIssueQty;
+	private float itemIssueQty;
 
-	private int itemRequestQty;
+	private float itemRequestQty;
 
-	private int itemPendingQty;
+	private float itemPendingQty;
 
 	private int status;
 	private int delStatus;
@@ -98,27 +98,27 @@ public class IssueReportDetail {
 		this.itemId = itemId;
 	}
 
-	public int getItemIssueQty() {
+	public float getItemIssueQty() {
 		return itemIssueQty;
 	}
 
-	public void setItemIssueQty(int itemIssueQty) {
+	public void setItemIssueQty(float itemIssueQty) {
 		this.itemIssueQty = itemIssueQty;
 	}
 
-	public int getItemRequestQty() {
+	public float getItemRequestQty() {
 		return itemRequestQty;
 	}
 
-	public void setItemRequestQty(int itemRequestQty) {
+	public void setItemRequestQty(float itemRequestQty) {
 		this.itemRequestQty = itemRequestQty;
 	}
 
-	public int getItemPendingQty() {
+	public float getItemPendingQty() {
 		return itemPendingQty;
 	}
 
-	public void setItemPendingQty(int itemPendingQty) {
+	public void setItemPendingQty(float itemPendingQty) {
 		this.itemPendingQty = itemPendingQty;
 	}
